Add: TimeIgnored for Ride

+++ UcabGo.Core/Data/User/Inputs/WalkingInput.cs
@@ -5,6 +5,6 @@ namespace UcabGo.Core.Data.User.Inputs
     public class WalkingInput : BaseRequest
     {
         [Required]
-        public float WalkingDistance { get; set; }
+        public int WalkingDistance { get; set; }
     }
 }
